UI fixes
Brought back black health bar separator - was moved to the right instead of the left.
Darkened health bar background.
Added second layer to the health heart, appears when the health bar moves past it.
Fixed issue that caused the timer to be slightly to the right - a space was added were it shouldn't be.
Fixed issue with card flash on selection not displaying properly after first selection.