Updated the stun mechanic on AI
Enabling stun now has a float variable for stun time, after which, the AI can recover.
There is also a bool for if the AI should wait until they have hit the ground to recover.
There is no game setting for this ATM as game settings has over gone a large overhaul recently.